Add files via upload
Early Spring School thrives in raising balanced children starting with the foundation of their education, which is why we practise full montessori and a delicate blend of the Nigerian and British curricula. Our focus as a school is to set a firm background for:

Academic Excellence; we prioritise academics, our curriculum is rigorous, well-rounded, and designed to provide our learners with a solid foundation of knowledge and skills. Our faculty comprises experienced and qualified educators who are passionate about teaching and committed to helping our students achieve their academic goals.